Merge pull request !264 from honglyua/sync_ixrt_model_zoo

+++ models/multimodal/diffusion_model/wan2.2-ti2v-5b/xdit/wan2.2_t2v_example.py
@@ -34,8 +34,8 @@ def main():
     pipe = xFuserWanPipeline.from_pretrained(
         pretrained_model_name_or_path=engine_config.model_config.model,
         engine_config=engine_config,
-        torch_dtype=torch.bfloat16,
-    )
+        dtype=torch.bfloat16,
+    ).to(torch.bfloat16)
 
     if engine_args.use_teacache:
         config = DiffusionCacheConfig(rel_l1_thresh = 0.2,
